Fix issue for comparison of same species
Cytoscape wasn't able to create nodes for one gene displayed in
different species because they would shar the same id. Now
for duplicate genes the network id is appended to the gene id,
the source and target of coexpression edges are changed
accordingly and orthology edges are duplicated for each network
of the same species.