Node highlighting enabled
As opposed to the old way of managing the extensions, the styling of the
nodes is controlled entirely by the content of the extension object in
the node data. If a node has the class 'extension', and the data
changes, the styling of the node should be updated. In addition to being
simpler to manage, it also means (or at least I think it means) that the
styling will be kept if the network is saved as JSON.